add scantailor-cli, abbyy, cancel feature

+++ diy/djos.py
@@ -2,6 +2,11 @@ import os, sys
 from threading import Thread
 from time import sleep
 from django.core import serializers
+import os
+import shutil
+from datetime import datetime
+from subprocess import call    
+
 
 def is_scantailor():
     ''' 
@@ -27,6 +32,50 @@ def is_scantailor():
             result = True
             
     return result
+
+def get_apps():
+    ''' 
+    this admittedly crude function checks and returns paths for popular diy book scanner software
+    '''
+    # can somebody fill in non-windows paths? 3-2-2012
+    result = {}
+
+    if sys.platform == 'win32': #Windows
+        apps = {'scantailor' :    ['C:/Program Files (x86)/Scan Tailor/scantailor.exe',
+                                   'C:/Program Files/Scan Tailor/scantailor.exe',
+                                  ],
+                'scantailor-cli': ['C:/Program Files (x86)/Scan Tailor/scantailor-cli.exe',
+                                   'C:/Program Files/Scan Tailor/scantailor-cli.exe',
+                                  ],
+                'abbyy':          ['C:/Program Files/ABBYY FineReader 5/finereader.exe',
+                                   'C:/Program Files/ABBYY FineReader 6/finereader.exe',
+                                   'C:/Program Files/ABBYY FineReader 7/finereader.exe',
+                                   'C:/Program Files/ABBYY FineReader 8/finereader.exe',
+                                   'C:/Program Files/ABBYY FineReader 9/finereader.exe',
+                                   'C:/Program Files/ABBYY FineReader 10/finereader.exe',
+                                   'C:/Program Files/ABBYY FineReader 11/finereader.exe',
+                                   'C:/Program Files (x86)/ABBYY FineReader 5/finereader.exe',
+                                   'C:/Program Files (x86)/ABBYY FineReader 6/finereader.exe',
+                                   'C:/Program Files (x86)/ABBYY FineReader 7/finereader.exe',
+                                   'C:/Program Files (x86)/ABBYY FineReader 8/finereader.exe',
+                                   'C:/Program Files (x86)/ABBYY FineReader 9/finereader.exe',
+                                   'C:/Program Files (x86)/ABBYY FineReader 10/finereader.exe',
+                                   'C:/Program Files (x86)/ABBYY FineReader 11/finereader.exe',
+                                  ]
+                }  
+        for app in apps:
+            for p in apps[app]:
+                if os.path.isfile(p):
+                    result[app] = p
+    elif sys.platform == 'darwin': # OsX
+        dirs = ['',]
+        result = None
+    elif sys.platform == 'linux2': #Ubuntu Linux
+        dirs = ['',]
+        result = None
+    
+    return result
+
 
 def get_drives():
     '''     
@@ -59,6 +108,8 @@ def get_mountpoint(request):
 
     if (delta):
         for drive in drives:
+            if os.system("cd " + drive + ":/100CANON") == 0: # try CHDK dir 
+                result = drive + ':\\100CANON\\'
             if os.system("cd " + drive + ":/DCIM") == 0: # try DCIM 
                 result = drive + ':\\DCIM\\'
             elif os.system("cd " + drive + ":/") == 0: # otherwise root
@@ -86,15 +137,16 @@ def create_project_dirs(cleaned_data):
             return False
     return True
 
-def import_pages(Page, Temp, src, dst, card='left'):
+def import_pages(p, Page, Temp, src, dst, card='left'):
     ''' copy and rename from card to computer '''
-    import os
-    import shutil
-    
     try:
+        t  = Temp(p='message',v='PLEASE WAIT -- Counting and enumerating your images...').save()
         files = os.listdir(src.replace('\\','/')) # eg 'e:/dcim'
     except:
         return '[{"error": "This directory does not exist."}]' # JSON syntax
+    
+    #delete entries for this card under this project before inserting below
+    pgs = Page.objects.filter(card=card,project=p).delete()
     
     types = ('jpg', 'jpeg', 'tif', 'tiff', 'png', 'jp2')
     inc_by = 2
@@ -108,10 +160,10 @@ def import_pages(Page, Temp, src, dst, card='left'):
     elif card == 'both':
         n=1
         inc_by = 1
-        
+    else:
+        return '[{"error":"Your card is not left right or both!"}]'
+    
     for item in files:
-        #p  = Page(card=card,m=total).save() -- TODO
-         
         i = item.split('.')
         ext = i[-1] # -1 returns last part
         if ext.lower() in types:
@@ -123,6 +175,7 @@ def import_pages(Page, Temp, src, dst, card='left'):
             n += inc_by
             output = "copied " + str(item) + " as " + renamed
             tasks.append([spath,dpath,output,task])
+            pg = Page(card=card,renamed=renamed,status='xfer and rename',xfer_date=datetime.now(),project=p).save()
     
     if not len(tasks):
         return '[{"error":"No images found!"}]'
@@ -130,8 +183,14 @@ def import_pages(Page, Temp, src, dst, card='left'):
     total = len(tasks) 
     i     = 1
     for spath, dpath, output, log in tasks:
+        # cancelled?
+        c = Temp.objects.filter(p='cancel')
+        if c:
+            t = Temp.objects.all().delete() 
+            return '[{"error":"Cancelled"}]'
+
         try:
-            t = Temp.objects.all().delete() # delete temp entries
+            t = Temp.objects.all().delete() 
             p  = round((float(i)/float(total))*100) # percent complete
             t  = Temp(p=output,k=i,v=p,m=total).save()
             shutil.copy2(spath, dpath)
@@ -140,24 +199,100 @@ def import_pages(Page, Temp, src, dst, card='left'):
                 sleep(1) # allows second thread to see 100 percent complete
             i += 1
         except:
-            return '[{"error":"An error occurred during copying and renaming"}]' # JSON syntax
+            return '[{"error":"An error occurred during copying and renaming. Verify your projects folder exists."}]' # JSON syntax
     t = Temp.objects.all().delete() # delete temp entries
     return '[{"success": "Import succeeded"}]' # JSON syntax
 
-def call_abbyy():
-    import os
-    from subprocess import call
+def run_batch(Temp, Page, p, path):
+    try:
+        bin = [get_apps()['scantailor-cli']]
+    except:
+        return '[{"error": "scantailor could not be found"}]' # JSON
     
-    f = []
-    for i in range(1,840):
-        n = '%(#)04d' % {'#':i} + '.jpg'
-        f.append(n)
+    path_in = path
+    path_out = os.path.join(path, 'scantailor')
+    flist = [] # file list
     
-    print f
+    #try creating dest
+    try:
+        os.makedirs(path_out, 0777)
+    except:
+        pass
     
-    call(['c:/program files/ABBYY FineReader 10/finereader.exe']+f,cwd='c:/img')
+    opts = ['-v']
+    opts.append('--layout=1')
+    opts.append('--dpi=400')
+    opts.append('--output-dpi=600')
+    opts.append('--color-mode=mixed')
+    opts.append('--white-margins=true')
+    opts.append('--normalize-illumination=true')
+    opts.append('--threshold=1')
+    opts.append('--match-layout-tolerance=1') # http://www.diybookscanner.org/forum/viewtopic.php?f=8&t=870&start=10
+    opts.append('--margins-left=5')
+    opts.append('--margins-bottom=5')
+    opts.append('--margins-left=10')
+    opts.append('--margins-right=10')
+    opts.append('--alignment-vertical=center')
+    opts.append('--alignment-horizontal=center')
+    
+    #options.append('--output-project=/somewhere') 
+    
+    pgs = Page.objects.filter(project=p).order_by('renamed')
+    flist = [os.path.splitext(x.renamed)[0] + os.extsep + 'tif' for x in pgs] # 0001.jpg becomes 0001.tif
+    
+    actions = ['OCR using ABBYY', 'ABBYY processed your file successfully.']
+    total = len(pgs) + len(actions)
+    i = 1
+    for pg in pgs:
+        # cancelled?
+        c = Temp.objects.filter(p='cancel')
+        if c:
+            t = Temp.objects.all().delete() 
+            return '[{"error":"Cancelled"}]'
 
-if __name__ == '__main__':
-    drive = get_usb()
-    print drive
+        f = [pg.renamed]
+        o = ['--orientation=' + pg.card]
+        
+        # update status in db 
+        pgs = Temp.objects.all().delete()
+        percent  = round((float(i)/float(total))*100) # progress
+        cmd = bin + opts + o + f + [path_out]
+        cmd_str = 'File: ' + f[0] + '<br/><br/>  <span id="progressbar-details-command"> Command: ' + ' '.join(cmd) + '</span><br/>'
+        t = Temp(p=cmd_str, k=i, v=total, m=percent).save()
+        
+        call(cmd,cwd=path_in)
+        i += 1
+    
+    # now abbyy
+    try:
+        bin = [get_apps()['abbyy']]
+    except:
+        return '[{"error": "abbyy could not be found"}]' 
+
+    opts        = ['/send','acrobat']
+    #opts.append('/send acrobat')  
+    cmd = bin + flist + opts
+    
+    #update progress
+    pgs = Temp.objects.all().delete()
+    percent  = round((float(i)/float(total))*100) # progress
+    t = Temp(p=actions.pop(0), k=i, v=total, m=percent).save()
+    
+    try:
+        call(cmd,cwd=path_out)
+    
+        #update progress    
+        i += 1
+        pgs = Temp.objects.all().delete()
+        percent  = round((float(i)/float(total))*100) # progress
+        t = Temp(p=actions.pop(0), k=i, v=total, m=percent).save()
+
+        return '[{"success": "Scantailor and ABBYY succeeded"}]'
+    except:
+        return '[{"error": "A problem occurred with ABBYY: bin=' + ' '.join(bin) + ' flist=' + ' '.join(flist) + ' opts=' + ' '.join(opts) +' cwd=' + ' '.join(path_out) + '"}]'
+    
+#if __name__ == '__main__':
+    #drive = get_usb()
+    #print drive
+
